refactor: extract note-ops handlers from App.svelte (#670, increment 4)
Move the note-lifecycle handler cluster out of App.svelte into
lib/app/note-ops.ts, and promote two pieces of shared UI state into runes
stores it (and other groups) depend on.

New stores:
- lib/stores/busy.svelte.ts — getBusyStore(): the modal busy-overlay label +
  withBusy() helper (was busyLabel + withBusy in App; used by ~19 handlers
  across note/source/refactor ops + the import-progress stream).
- lib/stores/clipboard.svelte.ts — getClipboardStore(): the multi-path
  cut/copy clipboard (was clipboardItems).

New module lib/app/note-ops.ts — createNoteOps(ctx) houses 16 handlers moved
VERBATIM: new note/folder, delete (+ safe-delete + executeDeletes +
open-first-reference), cut/copy/move/paste (+ clipboard summary), merge
(+ performMerge), rename, and prompt-driven copy/move. The only changes to the
moved bodies are dependency substitutions: component refs via ctx getters
(getSidebar / getEditorComponent), busy/clipboard via the new stores, and
safeDelete/mergePicker state via ctx setters. notebase/editor/api/dialog
stores + pure helpers resolve to module imports.

App builds the ctx (getters for the sidebar/editor refs so they're read live,
not captured) and destructures the bound handlers back into the SAME names, so
every call site (sidebar callbacks, command palette, keymap, native menu)
reads unchanged. Template now reads busy.label / clipboard.current reactively.

App.svelte: 3,451 → 2,965 lines (cumulative 3,764 → 2,965).

Verified: pnpm lint (0 errors), full suite (2778 passed), e2e boot smoke.
Reactivity reviewed by hand (runes stores; getters not cached; reactive
template reads). Follow-up: a behavioral unit test for createNoteOps would
net the moved bodies beyond the existing command/keymap dispatch tests.
